Added Renderer class, added new signals, added events for when new chapters, sections and pages are added.
The Renderer class basically takes the responsibility of rendering the
documentation for now it only provides a method to generate the index
for the entire docs as a multi-dimensional array.

Added the NEW_SECTION, NEW_CHAPTER, NEW_PAGE signals which are fired
when each is added.

ndoc will generate a book style index if run now which is pretty
freaking sweet!