Added settings files, removed Renderer, added Chapter generation.
The renderer class was removed as it was serving no purpose and only
doubling the indexing since this is already performed when ndoc parses
the diretory for contents. Added more settings, removing the theme
setting from the CLI option to a setting in the settings file. Chapter
generation has been moved into a event and moved outside of the main
Output class and the chapter generation now takes place inside of the
output generator being utilized. Index files are now parsed for
chapters using the chapter.php file.