Added h20 library courtesy of speedmax, previous and next chapter link generation.
Templates are now parsed using the h20 library allowing for django
style template syntax in the output themes. Previous and Next chapter
links generation is now available this works by transversing each next
available chapter and descending into each section within the chapter
until a new page is found, it attempts to descend each section depth to
the max in a section to allow for section hierarchy to take precedence.